+++ src/routes/api/auth/[...auth]/index.ts
@@ -1,3 +1,6 @@
-import { onRequest as onAuthRequest } from "~/server/auth";
+import type { RequestEvent } from "@builder.io/qwik-city";
 
-export const onRequest = onAuthRequest;
+export const onRequest = async (req: RequestEvent) => {
+  const { onRequest: onAuthRequest } = await import("~/server/auth");
+  return onAuthRequest(req);
+};
